move openssl part in its own file
