+++ lib/pem.js
@@ -1,17 +1,9 @@
 'use strict'
 
 var Buffer = require('safe-buffer').Buffer
-var spawn = require('child_process').spawn
-var pathlib = require('path')
-var fs = require('fs')
 var net = require('net')
-var crypto = require('crypto')
 var helper = require('./helper.js')
-var which = require('which')
-var osTmpdir = require('os-tmpdir')
-var pathOpenSSL
-var openSslVersion
-var tempDir = process.env.PEMJS_TMPDIR || osTmpdir()
+var openssl = require('./openssl.js')
 
 module.exports.createPrivateKey = createPrivateKey
 module.exports.createDhparam = createDhparam
@@ -69,7 +61,7 @@ function createPrivateKey (keyBitsize, options, callback) {
 
   params.push(keyBitsize)
 
-  execOpenSSL(params, 'RSA PRIVATE KEY', function (sslErr, key) {
+  openssl.exec(params, 'RSA PRIVATE KEY', function (sslErr, key) {
     function done (err) {
       if (err) {
         return callback(err)
@@ -104,7 +96,7 @@ function createDhparam (keyBitsize, callback) {
     keyBitsize
   ]
 
-  execOpenSSL(params, 'DH PARAMETERS', function (error, dhparam) {
+  openssl.exec(params, 'DH PARAMETERS', function (error, dhparam) {
     if (error) {
       return callback(error)
     }
@@ -136,7 +128,7 @@ function createEcparam (keyName, callback) {
     'explicit'
   ]
 
-  execOpenSSL(params, 'EC PARAMETERS', function (error, ecparam) {
+  openssl.exec(params, 'EC PARAMETERS', function (error, ecparam) {
     if (error) {
       return callback(error)
     }
@@ -245,7 +237,7 @@ function createCSR (options, callback) {
     helper.helperCreatePasswordFile({'cipher': '', 'password': options.clientKeyPassword, 'passType': 'in'}, params, delTempPWFiles[delTempPWFiles.length])
   }
 
-  execOpenSSL(params, 'CERTIFICATE REQUEST', tmpfiles, function (sslErr, data) {
+  openssl.exec(params, 'CERTIFICATE REQUEST', tmpfiles, function (sslErr, data) {
     function done (err) {
       if (err) {
         return callback(err)
@@ -373,7 +365,7 @@ function createCertificate (options, callback) {
     helper.helperCreatePasswordFile({'cipher': '', 'password': options.clientKeyPassword, 'passType': 'in'}, params, delTempPWFiles[delTempPWFiles.length])
   }
 
-  execOpenSSL(params, 'CERTIFICATE', tmpfiles, function (sslErr, data) {
+  openssl.exec(params, 'CERTIFICATE', tmpfiles, function (sslErr, data) {
     function done (err) {
       if (err) {
         return callback(err)
@@ -431,7 +423,7 @@ function getPublicKey (certificate, callback) {
     ]
   }
 
-  execOpenSSL(params, 'PUBLIC KEY', certificate, function (error, key) {
+  openssl.exec(params, 'PUBLIC KEY', certificate, function (error, key) {
     if (error) {
       return callback(error)
     }
@@ -464,7 +456,7 @@ function readCertificateInfo (certificate, callback) {
     '-in',
     '--TMPFILE--'
   ]
-  spawnWrapper(params, certificate, function (err, code, stdout) {
+  openssl.spawnWrapper(params, certificate, function (err, code, stdout) {
     if (err) {
       return callback(err)
     }
@@ -520,7 +512,7 @@ function getModulus (certificate, password, hash, callback) {
     helper.helperCreatePasswordFile({'cipher': '', 'password': password, 'passType': 'in'}, params, delTempPWFiles[delTempPWFiles.length])
   }
 
-  spawnWrapper(params, certificate, function (sslErr, code, stdout) {
+  openssl.spawnWrapper(params, certificate, function (sslErr, code, stdout) {
     function done (err) {
       if (err) {
         return callback(err)
@@ -556,7 +548,7 @@ function getDhparamInfo (dh, callback) {
     '--TMPFILE--'
   ]
 
-  spawnWrapper(params, dh, function (err, code, stdout) {
+  openssl.spawnWrapper(params, dh, function (err, code, stdout) {
     if (err) {
       return callback(err)
     }
@@ -592,9 +584,9 @@ function getDhparamInfo (dh, callback) {
  * @param {Object} options
  */
 function config (options) {
-  if (options.pathOpenSSL) {
-    pathOpenSSL = options.pathOpenSSL
-  }
+  Object.keys(options).forEach(function (k) {
+    openssl.set(k, options[k])
+  })
 }
 
 /**
@@ -620,7 +612,7 @@ function getFingerprint (certificate, hash, callback) {
     '-' + hash
   ]
 
-  spawnWrapper(params, certificate, function (err, code, stdout) {
+  openssl.spawnWrapper(params, certificate, function (err, code, stdout) {
     if (err) {
       return callback(err)
     }
@@ -676,7 +668,7 @@ function createPkcs12 (key, certificate, password, options, callback) {
     params.push('--TMPFILE--')
   }
 
-  execBinaryOpenSSL(params, tmpfiles, function (sslErr, pkcs12) {
+  openssl.execBinary(params, tmpfiles, function (sslErr, pkcs12) {
     function done (err) {
       if (err) {
         return callback(err)
@@ -718,7 +710,7 @@ function readPkcs12 (bufferOrPath, options, callback) {
     args.push('-nodes')
   }
 
-  execBinaryOpenSSL(args, tmpfiles, function (sslErr, stdout) {
+  openssl.execBinary(args, tmpfiles, function (sslErr, stdout) {
     function done (err) {
       var keybundle = {}
 
@@ -734,7 +726,7 @@ function readPkcs12 (bufferOrPath, options, callback) {
 
         if (keybundle.key) {
         // convert to RSA key
-          return execOpenSSL(['rsa', '-in', '--TMPFILE--'], 'RSA PRIVATE KEY', [keybundle.key], function (err, key) {
+          return openssl.exec(['rsa', '-in', '--TMPFILE--'], 'RSA PRIVATE KEY', [keybundle.key], function (err, key) {
             keybundle.key = key
 
             return callback(err, keybundle)
@@ -783,7 +775,7 @@ function checkCertificate (certificate, passphrase, callback) {
     helper.helperCreatePasswordFile({'cipher': '', 'password': passphrase, 'passType': 'in'}, params, delTempPWFiles[delTempPWFiles.length])
   }
 
-  spawnWrapper(params, certificate, function (sslErr, code, stdout) {
+  openssl.spawnWrapper(params, certificate, function (sslErr, code, stdout) {
     function done (err) {
       if (err) {
         return callback(err)
@@ -829,7 +821,7 @@ function checkPkcs12 (bufferOrPath, passphrase, callback) {
     args[3] = '--TMPFILE--'
   }
 
-  spawnWrapper(args, tmpfiles, function (sslErr, code, stdout, stderr) {
+  openssl.spawnWrapper(args, tmpfiles, function (sslErr, code, stdout, stderr) {
     function done (err) {
       if (err) {
         return callback(err)
@@ -865,7 +857,7 @@ function verifySigningChain (certificate, ca, callback) {
     '--TMPFILE--'
   ]
 
-  spawnWrapper(params, files, function (err, code, stdout) {
+  openssl.spawnWrapper(params, files, function (err, code, stdout) {
     if (err) {
       return callback(err)
     }
@@ -1166,197 +1158,3 @@ function readFromString (string, start, end) {
 
   return output
 }
-
-/**
- * Generically spawn openSSL, without processing the result
- *
- * @param {Array}        params   The parameters to pass to openssl
- * @param {Boolean}      binary   Output of openssl is binary or text
- * @param {Function}     callback Called with (error, exitCode, stdout, stderr)
- */
-function spawnOpenSSL (params, binary, callback) {
-  var pathBin = pathOpenSSL || process.env.OPENSSL_BIN || 'openssl'
-
-  testOpenSSLPath(pathBin, function (err) {
-    if (err) {
-      return callback(err)
-    }
-    var openssl = spawn(pathBin, params)
-    var stderr = ''
-
-    var stdout = (binary ? new Buffer(0) : '')
-    openssl.stdout.on('data', function (data) {
-      if (!binary) {
-        stdout += (data || '').toString('binary')
-      } else {
-        stdout = Buffer.concat([stdout, data])
-      }
-    })
-
-    openssl.stderr.on('data', function (data) {
-      stderr += (data || '').toString('binary')
-    })
-    // We need both the return code and access to all of stdout.  Stdout isn't
-    // *really* available until the close event fires; the timing nuance was
-    // making this fail periodically.
-    var needed = 2 // wait for both exit and close.
-    var code = -1
-    var finished = false
-    var done = function (err) {
-      if (finished) {
-        return
-      }
-
-      if (err) {
-        finished = true
-        return callback(err)
-      }
-
-      if (--needed < 1) {
-        finished = true
-        if (code) {
-          if (code === 2 && stderr === '') {
-            return callback(null, code, stdout, stderr)
-          }
-          return callback(new Error('Invalid openssl exit code: ' + code + '\n% openssl ' + params.join(' ') + '\n' + stderr), code)
-        } else {
-          return callback(null, code, stdout, stderr)
-        }
-      }
-    }
-
-    openssl.on('error', done)
-
-    openssl.on('exit', function (ret) {
-      code = ret
-      done()
-    })
-
-    openssl.on('close', function () {
-      stdout = (binary ? stdout : new Buffer(stdout, 'binary').toString('utf-8'))
-      stderr = new Buffer(stderr, 'binary').toString('utf-8')
-      done()
-    })
-  })
-}
-
-function spawnWrapper (params, tmpfiles, binary, callback) {
-  if (!callback && typeof binary === 'function') {
-    callback = binary
-    binary = false
-  }
-
-  var files = []
-  var delTempPWFiles = []
-
-  if (tmpfiles) {
-    tmpfiles = [].concat(tmpfiles || [])
-    params.forEach(function (value, i) {
-      var fpath
-      if (value === '--TMPFILE--') {
-        fpath = pathlib.join(tempDir, crypto.randomBytes(20).toString('hex'))
-        files.push({
-          path: fpath,
-          contents: tmpfiles.shift()
-        })
-        params[i] = fpath
-        delTempPWFiles[delTempPWFiles.length] = fpath
-      }
-    })
-  }
-
-  // TODO: need to refactored
-  files.forEach(function (file) {
-    fs.writeFileSync(file.path, file.contents)
-  })
-
-  spawnOpenSSL(params, binary, function (err, code, stdout, stderr) {
-    helper.helperDeleteTempFiles(delTempPWFiles, function (fsErr) {
-      callback(err || fsErr, code, stdout, stderr)
-    })
-  })
-}
-
-/**
- * Spawn an openssl command
- */
-function execOpenSSL (params, searchStr, tmpfiles, callback) {
-  if (!callback && typeof tmpfiles === 'function') {
-    callback = tmpfiles
-    tmpfiles = false
-  }
-
-  spawnWrapper(params, tmpfiles, function (err, code, stdout, stderr) {
-    var start, end
-
-    if (err) {
-      return callback(err)
-    }
-
-    if ((start = stdout.match(new RegExp('\\-+BEGIN ' + searchStr + '\\-+$', 'm')))) {
-      start = start.index
-    } else {
-      start = -1
-    }
-
-    // To get the full EC key with parameters and private key
-    if (searchStr === 'EC PARAMETERS') {
-      searchStr = 'EC PRIVATE KEY'
-    }
-
-    if ((end = stdout.match(new RegExp('^\\-+END ' + searchStr + '\\-+', 'm')))) {
-      end = end.index + (end[0] || '').length
-    } else {
-      end = -1
-    }
-
-    if (start >= 0 && end >= 0) {
-      return callback(null, stdout.substring(start, end))
-    } else {
-      return callback(new Error(searchStr + ' not found from openssl output:\n---stdout---\n' + stdout + '\n---stderr---\n' + stderr + '\ncode: ' + code))
-    }
-  })
-}
-
-/**
- *  Spawn an openssl command and get binary output
- **/
-function execBinaryOpenSSL (params, tmpfiles, callback) {
-  if (!callback && typeof tmpfiles === 'function') {
-    callback = tmpfiles
-    tmpfiles = false
-  }
-  spawnWrapper(params, tmpfiles, true, function (err, code, stdout, stderr) {
-    if (err) {
-      return callback(err)
-    }
-    return callback(null, stdout)
-  })
-}
-
-/**
- * Validates the pathBin for the openssl command.
- *
- * @param {String} pathBin The path to OpenSSL Bin
- * @param {Function} callback Callback function with an error object
- */
-function testOpenSSLPath (pathBin, callback) {
-  which(pathBin, function (error) {
-    if (error) {
-      return callback(new Error('Could not find openssl on your system on this path: ' + pathBin))
-    }
-
-    callback()
-  })
-}
-
-/**
-* Once PEM is imported, the openSslVersion is set with this function.
-*/
-spawnOpenSSL(['version'], false, function (err, code, stdout, stderr) {
-  var text = String(stdout) + '\n' + String(stderr) + '\n' + String(err)
-  var tmp = text.match(/^LibreSSL/i)
-  openSslVersion = tmp && tmp[0] ? 'LibreSSL' : 'openssl'
-
-  openSslVersion = openSslVersion.toUpperCase()
-})
