+++ package.json
@@ -7,6 +7,7 @@
       "url": "https://www.josef-froehle.de/"
     }
   ],
+  "license": "MIT",
   "name": "pem",
   "description": "Create private keys and certificates with node.js and io.js",
   "version": "0.0.0-development",
